rhi: Fix using the same srb with different pipelines
...within the same frame. Applies to D3D11, D3D12, and Metal.

The problem is ultimately unique to backends that simulate the command
buffer (so D3D11, OpenGL) and the command buffer entry added from
setShaderResources() references the srb object. And it is only really a
problem if the QRhiShaderResourceBindings implementation has "state",
and the processing of that command buffer entry accesses that state.
This leaves D3D11 only, in the GL case there's no state in the srb. This
issue is then intertwined with another problem, again unique to the
native resource mapping tables used for HLSL shaders, namely that a
different shader may use a different native HLSL register for a SPIR-V
style binding point in the QShaderResourceBindings, even though the srb
is otherwise fully fine for both shaders.

Qt Quick or Quick 3D do not typically exhibit command sequences such as:

setGraphicsPipeline(pipeline_a);
setShaderResources(srb);
draw
setGraphicsPipeline(pipeline_b);
setShaderResources(srb); // the same one
draw

However, it is perfectly valid to have different pipelines with shaders
that have the same (or a subset) resources (uniform buffers, textures),
and use the same srb object with all these pipelines, if the shaders are
expected to see the exact same data.

This is not always causing a problem with the D3D11 backend. However, if
pipeline_a and pipeline_b are using different shaders, with pipeline_b
perhaps using a uniform buffer that pipeline_a does not (but the same
srb is still just fine), that leads to erroneous (e.g., swapped) buffer
bindings and the the HLSL shader reading bogus data for some uniforms.

First, setShaderResources() needs to recognize (with D3D11) that the
graphics pipeline is now different, since the check that goes through
the binding list in the srb to see if the referenced buffers and
textures have changed is not sufficient if nothing has changed but the
pipeline is now different (which may involve different resource mapping
tables for the shaders!).

Then, solving this breaks the first draw call in the above example
sequence instead of the second, whereas originally it was the second
draw call that was rendering incorrectly, not the first one.

This is due the srb being referenced and the internal data accessed when
executing the command buffer. The data there is for the second draw call
(setShaderResources) at that point! (and not the first, the state
generated in the first setShaderResources is lost)

Solve this by putting the necessary state (the prepared resource and
other lists that will be passed to VSSetConstantBuffers1 and the
countless other similar resource setting functions of D3D11) into (or
next to) the command buffer, so the second setShaderResources() with the
same srb does not break havoc and mess up the first one.

Third problem: when adding an autotest for this (or parts of the
problem at least), it turns out D3D12 and Meta are problematic too,
even though the problem in the bug report is not reproducable with
those. The autotest shows however that the D3D12 backend also needs
the "different pipeline should trigger things even if the srb is the
same" fix.

Fixes: QTBUG-140795
Pick-to: 6.9 6.8
Change-Id: Ife7703cdda67209500df52c5a46a92a1017bbc44
Reviewed-by: Kaj Grönholm <kaj.gronholm@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 578e8a01a23ab13f382bcfb339dc8309269940d7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>