QWidgetAnimator: make sure to remove all destroyed animations
Animations were only removed from the animation map when they were
finished but not when they were destroyed before. In this case the
QPointer was reset to a nullptr but might dereferenced later when a new
animation on the related widget was started. A change from raw pointer
to QPointer was added a long time ago to fix a spurious crash in
QDockWidget handling but this was not enough. Interestingly it was not
noticed for over 12 years...
This amends 33214af3784feacb2d5188bbf07da92f45f582f9

Pick-to: 6.8 6.5
Fixes: QTBUG-140207
Change-Id: I9f0d10bf7356d00cfba232ae2f3c3a7704b27cef
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
(cherry picked from commit 659a100f483ce706c6ea60dc0db6331f7c2b694b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>