Create infrastructure for variance calculations in benchmarks
Create the needed stubs for calculating variance in benchmark
operations. Add a variance field to QBenchmarkMeasurerBase::
Measurement struct. Give zero value to all measurements for
this new field. Add no-op virtuals to QBenchmarkMeasurerBases
to be gradually overloaded in due course.

[ChangeLog][QtTest] Add reporting of variance or standard deviation
to benchmark metrics, whenever more than one run of the measurement
contributed to the result, if supported by the measurer.

Task-number: QTQAINFRA-7315
Change-Id: Icfca0b9027418218cf035390ea46cb9a088e28ae
Reviewed-by: Jason McDonald <macadder1@gmail.com>