QWizard: make QWizardHeader a bit more safe
The Ruler ctor is only used from QWizardRuler, a subclass, so mark it
as protected, because it doesn't initialize any of the data members.

Found while fixing a Coverity complaint in the vicinity.

Amends the start of the public history.

Pick-to: 6.8 6.5
Change-Id: I407940dc8f720976dec0608eae1649d088f6f98d
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 38a9db9f63a4ebdff86cdf726ea50e55413bcedf)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>