Testlib: disable App Nap for whole execution of the test
The intention was to disable App Nap for the whole test. However it was
a RAII class which was destroyed at the end of QTest::qInit. This patch
moves the AppNap disabler into a global variable, which is cleared in
qCleanup.

Pick-to: 6.8
Change-Id: I5c267b2b66626b230851e61e79655ef3a30d4481
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
(cherry picked from commit 305e1ba820ff02d8aa46e902634590c307c21f17)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>