Fix getting color scheme in x-d-p platform theme
Commit 4ac89da broke reading system color scheme support from the
settings portal of xdg-desktop-portal. The cause is that the metatypes
provided for the parameter and result of the ReadAll method are incorrect and don't match the declared interface which is:

ReadAll (
  IN namespaces as,
  OUT value a{sa{sv}}
)

This patch ensures that the parameter and return types used with QDBus
are correct, and adds warning logs in case of future failures.

Change-Id: I6ac51dbfa1e9f1ebcb59e8398b6c5119da79d1f8
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
(cherry picked from commit 9850c465c900bfc0b43c4ab21e3c361aa39ff417)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>