QRM: don't return pointer-to-const from parentRow
We want to be able to get the parent from one row and make it the parent
of another row, and returning a `const row *` from parentRow makes that
impossible without a hacky const-cast.

Returning a pointer to const is not necessary. The inner state of the
parent row has no impact on the observable state of the child from which
it is returned.

Change-Id: I1487cf2a2ed5122dd0e3530af6e055ed56e34fb7
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 72cee72eb0913d049e352a7b45920e6012df1aa6)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>