macOS: Apply icon tint color without intermediate drawing step
As part of 9f392c09a1d30e48494b4df0e2f5f531c7e4ec4b we moved from
requesting hierarchical icons to automatic rendering mode, as the
hierarchical mode with a single tint color produced a different look
than the default look of the system, which was primarily monochrome.
We did so by drawing the image in an intermediate step with the tint.

But there is a better way. If we first create a symbol configuration
with a single palette color, which produces a configuration with palette
rendering mode (which we don't want, as that also has a different look),
and then apply a configuration with an explicit request for monochrome
icons, we get the tint with the exact same look as the system.

This keeps the NSImage as a symbol image internally for as long as
possible, instead of flattening it down to a pixmap during configure.

Pick-to: 6.8
Change-Id: Ifad9b452b54cb6862396eb0fa6df9f447397a405
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 7c996f4d66e7c2fac0bcb6d8508e49c12464e463)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>