Wayland: handle delayed CSD creation following decoration negotiation
When window decoration negotiation takes place such that it results in
the zxdg_toplevel_decoration_v1.configure event telling that the client
should do CSDs coming slightly before or together with the first
xdg_toplevel.configure event - the decorations are created during the
handling of the latter. But that processing relies on geometry data
from before creation of the decorations, which is stale.

Handle this by moving the call to QWaylandWindow::windowContentGeometry
to be after the call to handleWindowStatesChanged, since the latter can
cause changes to the gemoetry of the window.

Fixes: QTBUG-140847
Change-Id: I5157f7d7f2fd4f45aaf7c8209cded565bd7cd21e
Reviewed-by: Ilya Fedin <fedin-ilja2010@ya.ru>
Reviewed-by: David Edmundson <davidedmundson@kde.org>
(cherry picked from commit 8fdf90174797bf1a67ac2d91aedee7a18400489f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>