windeployqt: Enable deployment of compiler runtime for mingw/clang
The code in question was present but the user was unable to use the
option. As the code basically works the same as for other configurations
(namely mingw g++) the option is now enabled by default.

The support for mingw/clang was added in
3d0fb2cea9c4e392497f9067d21b194124de9333 and enabling the code path was
just forgotten.

Fixes: QTBUG-140789
Change-Id: I9709a49ffde4a93c3cf6a8b9c0b1eea20cdefc1f
Reviewed-by: Wladimir Leuschner <wladimir.leuschner@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
(cherry picked from commit ae88ad1a1569153160d7998346b8dbe03cbdfa76)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>