Fix crash in QWidgetWindow::handleMouseEvent
User code that calls
    qApp->processEvents();
in context menu actions could lead to accesses of deleted memory.

The reason was that the QWidgetWindow object was deleted as an effect of
the event delivery in handleMouseEvents, and later accesses to object
members in this method crashed.

Pick-to: 6.8
Fixes: QTBUG-138419
Fixes: QTBUG-140132
Change-Id: Ia61e4380b29333875de9e1202c363d99d3f79e2a
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit b92724a1be2475dac052308e5c75e7bf52813fb8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>