qt-testrunner: consider exit codes >=128 as CRASH
Previously it was only < 0, which was valid because Python set those on
POSIX platforms to represent signal number. By checking >=128 we also
accommodate for crashes on windows and other unpredictable situations
(since QTest is limiting itself to [0,127]).

Also change the exit code of the mock test used for qt_testrunner
testing. 130 is what is shown by bash when a process is interrupted by
Ctrl-C (SIGINT), while 131 is what is shown for SIGQUIT and does not
happen that easily. Completely insignificant change for the python
programs like qt-testrunner and its tests, since subprocess.run()
returncode is negative in case of a signal. Only visible if qt_mock_test
is invoked in the shell for debugging purposes.

Pick-to: 6.10 6.8
Change-Id: I4759bc0e74ca2fb1d1594c71e541feb5363bb7ea
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
(cherry picked from commit fa384e4b8f2a5ae7289e7eb11b77e330bcb31457)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>