Fix artifacts on distance fields with overlapping lines
The latest version of Noto Sans CJK JP has some very small
bugs where a tiny line segments extend from the outline and then
return immediately with an overlapping line (e.g. one line with
a normal of (0, 1) first and then another line with a normal of
(0, -1) right after it).

The way distance field generation works is that it will first
flatten curves into line segments. Then it will gradient fill
rectangles around these line segments, aligned with the normals.

If there is an angle between the line segments, there will be
gaps between these rectangles, so we then need to fill these in
with triangles.

There are two different code paths for the triangles: One if
the angle between the two line segments is convex and one if
not.

For parallel lines, there is no gap to fill, but we would still
execute the triangle pass for concave in this case, as parallel
lines were considered non-convex. This broke down for these
overlapping lines, causing large smudges of fully interior
pixels to be drawn over the position of the overlapping lines.

Since we know that parallel lines will not require the fill
triangles, we can bypass this step entirely if the line segments
are parallel, which eliminates this issue.

Fixes: QTBUG-139610
Change-Id: I18c5480630ae667f511843881b1243628d9a17c5
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
(cherry picked from commit ad7a9155c1b855ed6e06929fb283254910059231)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>