Fix QLineEdit frame color when using gradient style sheet and Fusion
When using a  gradient style QBrush via style sheet, an
invalid color is set on the widget palette for the Window role.

This color is used to in QFusionStylePrivate::outline() to calculate
the outline color, resulting in a black/invalid color as well for
PE_FrameLineEdit. Fall back to the widget palette in case an invalid
color is encountered.

Fixes: QTBUG-138960
Change-Id: Id1b5320476e59a743dfd0cf489032fe34e47d422
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 439b758811486cbc75419df92192ef5307952ade)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>