QRM: don't overwrite table row when setting cell with RangeModelDataRole
A list of multi-column gadgets or objects becomes a table, and writing
Qt::RangeModelDataRole to any cell should overwrite that property, and
not the entire gadget/object for the whole row.

Add test case that fails with an attempt to replace the Object pointer
with a string value.

Change-Id: I1aded3be7c9a33e2912010c011047c887a6f2354
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit ad993827301206b9d179f35b88cbf8621b008853)
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>