tst_QSettings: split the testEscapes() test: testEscapedVariant
Making it easier to debug.

The original code explicitly compared the input QVariant with
the variant-converted-to-the-underlying-type, most likely to
test the conversion round-trip. This is already covered by
QSettingsPrivate::{variantToString,stringToVariant}() which
does the conversion internally.

Use example.org instead of software.org, the latter is regitered
and we shouldn't use it in tests.

Pick-to: 6.8 6.5
Change-Id: Ic866f39af50370079bcfefc24c9093d733e32f9e
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
(cherry picked from commit ef686ad94ab0f5371e24440c5cb50b469a0ce86c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>