Windows: Fix QWindowWindows::requestUpdate
It's very possible for the platform window to be released and recreated
for a QWindow. We therefore still need to verify that we both have
a platform window, and that it's the same window we triggered the
update for.

This change amends 31c38d5f766366931cd127f7323f9c26ac5de498

Fixes: QTBUG-141128
Change-Id: Id6e3fcbe3c6d2d5aa30b2af256a74de2fc9e5f9e
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
(cherry picked from commit f3e029719b14bf2180d247afebaceea3607a323c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>