QApplication: reset WA_KeyboardFocusChange on focus change
WA_KeyboardFocusChange is set when the user changes the focus by mouse
but never cleared once the focus was changed e.g. with the mouse. The
only way to clear it is to unfocus the whole top-level widget. This is
results in focus rects in some style like fusion on checkboxes where
there shouldn't be one.
Fix it by clearing the attribute when the FocusReason is not keyboard.

Fixes: QTBUG-141100
Change-Id: I211e7f435c9fe748ee9925e35b481166b7446a55
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit 6cd906a8bd80b62d254c9491720909cd87cfa36b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>