QFuture::cancelChain(): improve propagation through unwrap()
QFuture::unwrap() returns a new QFuture that is not added to
the continuation chain, but potentially fulfilled by both outer
and nested futures.
As a result, the parent-child relation in the continuation chain
is broken, and a cancelChain() call will not be propagated up
the chain through unwrap().

Fix it by explicitly adding the newly-created QFutureInterface
into the hierarchy. This requires using some private API of
QFutureInterfaceBase, so add UnwrapHandler as a friend class.

There is still one more pending issue with the cancellation
propagation: if at the time of cancellation the outer future is
already finised and the nested is in progress, the cancellation
will not be propagated into it, and so it will finish successfully.
This issue should be addressed separately.

Task-number: QTBUG-140786
Pick-to: 6.10
Change-Id: I62a4038f4544fff3cca65cb817940ae8b48db384
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
(cherry picked from commit ef399385253c93b3fbe81251ef71aa47de4ad016)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>