QStringIterator: add nextOrRawCodeUnit()
While fixing Clang 21 -Wcharacter-conversion warnings, I came across
many code snippets that used a pattern like

    char32_t ucs4 = string[i];
    if (QChar::isHighSurrogate(ucs4) && i + 1 != len) {
        ushort low = string[i + 1];
        if (QChar::isLowSurrogate(low)) {
            ucs4 = QChar::surrogateToUcs4(ucs4, low);
            ++i;
        }
    }

(this one from qunicodetools.cpp)

The natural question is why this was never ported to QStringIterator,
and the answer is: because QStringIterator doesn't support this.

Add new functions nextOrRawCodeUnit() (and previousOrRawCodeUnit()),
to fix this shortcoming.

The name of the functions is chosen so that it will work also in a
future UTF-8 string iterator.

Add tests that verify that the old code and the new functions produce
the same series of code points.

Pick-to: 6.8 6.5
Change-Id: I34ba8e416ee290badc1c16e33b46a17d56741574
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
(cherry picked from commit bb7f82a410971bc73e2ac4d10cf6c285ddee9e42)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>