tst_QList: add and test NoexceptMovable
This is what Movable should have been: the move constructor and
assignment operators are noexcept (the copy ones are too, but they
didn't need to be). I'm not modifying the existing class because QList
has behaviors that differ if the copy/move construction can throw, so
it's best to test them both.

Pick-to: 6.8 6.5
Task-number: QTBUG-141366
Change-Id: I00d7f390dc46b35ee78afffd1b591c8972280493
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 026c563ee0422c7476992457955fc6529ddc9667)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>