QTZP: add findOffsetPrefix() helper for parsing localized zone offsets
Initial implementation is painfully complex and expensive, inferring
the format of offsets by asking for the suitably-formatted names of
chosen-offset zones so as to match the text to be parsed against these
and infer which parts of the text thus correspond to hour, minute and,
if ever relevant, seconds. Individual backends may, in due course,
offer more efficient solutions.

Task-number: QTBUG-139223
Change-Id: Iba9515e76371ead32e7cfe7be5a15db49e7e2c9c
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>