Tests: qhash - clear global variable use in stdHashImpl
`StdHashUsed` is set in the stdHashImpl, but the test does not ensure
that it is initialized before the test. This causes test failures with
-repeat.
Clearing the flag at the beginning of the test function fixes this
behaviour.

Change-Id: I383611ce9368c6826cb5436a74e12bd73748491c
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
(cherry picked from commit c04f85e4e02e2fb99d7fe8c766a097c4d9cf8083)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>