qt-testrunner: check validity of the XML file on test function re-runs
Tests on certain platforms run with specific wrappers, that might have
trouble reporting back the proper exit code.

For example, on Android each test is transfered to the device/emulator
and is executed not as a separate process, but as an activity. The
equivalent of "exit code" is caught and returned by a special wrapper
script for the platform.

It happens sometimes that these wrapper scripts fail to report back
correctly, and report zero (0) despite failed tests. For that reason we
now parse the test XML log on individual test re-runs too, and reporting
inconsistencies as CRASH, like we do with the main test execution.

Task-number: QTQAINFRA-7349
Task-number: QTQAINFRA-7378
Pick-to: 6.10 6.8
Change-Id: I27525f22331d44141be8825786a6f71e89543e92
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
(cherry picked from commit 421059ce08dc130f615ca1404e48a09d6936acb3)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>