CUPS print dialog: Support entering custom values for PPD options
CUPS not only supports PPD options that have a fixed
set of values to choose from, but also PPD options
that allow entering a custom value.

CUPS documentation on custom options: [1]

Implement support for such custom PPD options by
adding a QLineEdit to enter custom text when a non-null
return value of ppdFindCustomOption [2] indicates that
a custom option exists.

Show that one only if the "Custom" value has been
selected for the PPD option. In that case, set
manually entered value in the print options
in QPrintPropertiesDialog::setPrinterAdvancedCupsOptions
as "Custom.<entered_value>" which matches the behavior
passed of other CUPS client application implementations
like GTK 3 applications or LibreOffice.

This commit implements passing a value manually entered
in the Qt print dialog to the CUPS print job.

So far, responsibility that the user-entered text
meets the type and minimum/maximum length supported
for the custom option is fully on the user, i.e.
the value is simply passed to CUPS as is.

[1] https://www.cups.org/doc/spec-ppd.html#OPTIONS
[2] https://www.cups.org/doc/api-ppd.html#ppdFindCustomOption

Task-number: QTBUG-137426
Pick-to: 6.12 6.11 6.8
Change-Id: Ia0d9260170a691f1e42a4e611eed86759e97dcdf
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>