CUPS print dialog: Set custom option's default value
Add a new QVariant parameter to the QPrintDevice::property
and QPlatformPrintDevice::property methods and
make use of that to implement logic to get the
value for a PPD option in QPpdPrintDevice::property.

In the print dialog, make use of that to retrieve the
(default) value of custom PPD options [1] and set that
for the line edit that allows setting a custom value.

[1] https://www.cups.org/doc/spec-ppd.html#OPTIONS

Fixes: QTBUG-137426
Change-Id: Ibff198178b34975cfa5f49ffd9cdd17cb3e46897
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>