Make QTZP::findLongNamePrefix() cope with zones unknown to CLDR
The backend may know about zones for which our most recent CLDR update
lacked data, for example because one has been added to the IANA
database recently, as has happened for America/Coyhaique in release
2025b. Because findLongNamePrefix() was working with a local sui
generis type to hold details by which to look up an IANA ID in the
CLDR tables, if it found a zone that matched for the locale's region
format, it had to find the index of that zone in the tables to fill in
the sui generis type's field, even though that was only used to look
up the IANA ID later. That lookup failed if the IANA ID wasn't in the
CLDR tables. So switch to using a NamePrefixMatch to hold details in
the final region format search, so that we can simply store the ID
rather than an index that might not exist.

Add the new zone that triggered this to test data and fix alphabetic
order in its list while I'm about it.

On picking to 6.10:
* Some offset-based matches are attempted after the region format
  search, so continue with those if it found nothing.
* Tests aren't split between QTZ and its backend, so apply to QTZ
  tests the changes originally applied to backend tests. These lack
  some of the zones added in more recent changes, so settle for fixing
  up alphabetic order and making the same one addition.

Amends commit 2edd9286cf386675be76032424248e60216f6331 (6.10).

Change-Id: I7c4850f8d368460852c7d19801ee80cf25ccf7e8
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
(cherry picked from commit 571895b09dd2e12b3bd30658656d4ec4c5f36878)
Reviewed-by: Mate Barany <mate.barany@qt.io>