CMake: Improve SBOM relationship handling for targets
Previously custom relationship information could be specified for
targets via the SBOM_RELATIONSHIPS option to _qt_internal_add_sbom
or qt_internal_extend_target.

Unfortunately that option expected unstructured spdx v2 strings, which
prevents us from using it for cyclonedx or a future spdx v3 API.

Introduce a new option called SBOM_RELATIONSHIP_ENTRIES that takes
structured values similar to how we do it for PURL entries.

An example:

_qt_internal_extend_sbom(${target}
    SBOM_RELATIONSHIP_ENTRIES
        SBOM_RELATIONSHIP_ENTRY
            SBOM_RELATIONSHIP_FROM
                CustomLinkerTarget
            SBOM_RELATIONSHIP_TYPE
                BUILD_DEPENDENCY_OF
            SBOM_RELATIONSHIP_TO
                Qt::Core
            SBOM_RELATIONSHIP_COMMENT
                "Linker used to link QtCore library"
)

Notably the FROM and TO options expect target names enriched with SBOM
information with _qt_internal_add_sbom(), not bare spdx ids or
cdx bom-refs. A previous iteration attempted to create an API based
on such ids, but it makes the API more complicated due to differences
in how spdx and cyclonedx handle external references.

A future change will address creation of custom relationships for
external referenced targets as well.

Change the code to assemble the relationship info from entries such
as the one above, together with relationships created based on target
linkage.

Adapt the SPDX v2 serialization code to ensure everything validates
properly. Also adapt the CycloneDX serialization script.

Because CycloneDX is a bit more limited in what relationship info it
can represent, we only extract the "from" and "to" parts for it,
discarding the comment and relationship type. We still write out the
information into the intermediary toml file because it might be useful
in the future.

The relationship type should be one of the strings listed in the spdx
v2 spec. See
https://spdx.github.io/spdx-spec/v2.3/relationships-between-SPDX-elements/

Keep the old SBOM_RELATIONSHIPS option for backward compatibility,
because WebEngine uses it and has not yet been migrated to the new
format.

Added test cases to verify the new feature.
Additional test infrastructure code is added to try and validate
the existence of the relationships in the generated files. That code
can probably be improved, but it gets the job done for now.

Pick-to: 6.10 6.8
Task-number: QTBUG-142899
Change-Id: I7230b0711384d194249d50a06f806641305704e0
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
(cherry picked from commit f052e67a0d4f1ab849b4e3b54779b7b19df7416b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>