qHash: don't hash the padding bytes of long double
On x86, the x87 FSTP instruction only writes 10 bytes, in spite of the
long double type being 16 bytes (12 on 32-bit). That means we left some
extra bytes uninitialized and used the in the in the murmurhash()
calculation.

  117a40:       fldt   0x10(%rsp)
  117a44:       lea    0x10(%rsp),%rdi
  117a49:       fldz
  117a4b:       faddp  %st,%st(1)
  117a4d:       fstpt  0x10(%rsp)
  117a51:       call   115fa0 <murmurhash>

This also means the test never failed in the CI or for a developer
because those extra bits happened to consistently be the same, a fragile
situation.

Simply initializing the buffer isn't enough, because sometimes the
compiler will write all bytes: for example, in a debug build, it often
spills the 10 bytes to the stack, then memcpy()s the 16 bytes to our
buffer (it simply copies uninitialized bytes over). Therefore, we
recalculate how many bytes the type actually occupies from the
numeric_limits information. I've used the same calculation elsewhere for
FP16 and I've verified it works for FP128 too (GCC's manual suggests
it's the default for Android/Bionic) - I won't guarantee it works for
HF8 an BF8.

For the test, I've opted to simply make use of the stack, and force all
buffers to be spilled by calling a function that cannot be inlined. I
can't guarantee that useStack() uses stack where the helper later will.
That's a bit fragile.

This does not apply to MSVC, because Microsoft dropped support for the
extended precision there (long double and double are the same).

[ChangeLog][QtCore][QHash] Fixed a bug that caused the qHash() function
to return non-deterministic results for long double values on most x86
systems, regardless of seed.

Fixes: QTBUG-141637
Pick-to: 6.10 6.8 6.5
Change-Id: Iefa432aceadcd8b30391fffdd292c557da6eee22
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
(cherry picked from commit ee4d239ee5b13b8dad925a51ff1db0a3728a4675)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>