a11y Android: Inform TalkBack about used language
This is helpful if the app's language is different from the OS language.
Without this change english texts are read with a german accent if
the OS language is set to German and the app language to English.

For now, only QAccessibleWidget, QAccessibleQuickControl, and
QAccessibleQuickItem return useful values for Attribute::Locale.
All three rely on specific setter (QWidget::setLocale,
Control.locale, ApplicationWindow.locale for the QML Item) and
do not consider other application wide means of setting the
language (like installed QTranslator or QJSEngine::uiLanguage).

Task-number: QTBUG-141701
Pick-to: 6.11
Change-Id: Id4bf0280469229c202ab3cf2480903212d351c2b
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>