Doc: Collect disabled QDoc warnings to a dedicated qdocconf file
A number of QDoc warnings are ignored for Qt module documentation
builds. Add them to a dedicated configuration file, included from
compat.qdocconf.

Add 'Undocumented property' warning, introduced to QDoc on 6.11,
to the list of ignored warnings for 6.10 and earlier. This is to
avoid integration issues as documentation testing on all active
branches uses QDoc from (or close to) dev/HEAD.

Pick-to: 6.8 6.5
Task-number: QTBUG-140629
Change-Id: I6fbe8a54e06e4d03221bc41bc63c0e9697b313d5
Reviewed-by: Paul Wicking <paul.wicking@qt.io>