Stop asserting after receiving an HTTP 2 frame with the wrong size
Skip the body of a frame when the frame has the wrong size. If
the entire body cannot be skipped, continue skipping later.

This is needed when receiving frames of wrong size of types DATA,
PRIORITY, or WINDOW_UPDATE. For other frame types and when the
stream id is 0: a GOAWAY frame is sent and the connection is
closed. But for these three types an RST_STREAM frame is sent,
and after that reading frames from the socket continues.

Pick-to: 6.10 6.8 6.5
Fixes: QTBUG-141207
Change-Id: Ib823da16d0c522768ff272f34b8dd56255877732
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
(cherry picked from commit a0147341cb8b660879235c10414c3b78f45dee48)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>