QComboxBox: replace className() comparison with qobject_cast
The old code checked for a variable's dynamic type by calling its
metaObject()->className(), comparing it to a dynamically-allocated
QByteArray.

Instead of fixing just the allocation, let's use a compile-time
checked construct: Unlike inherits() and className(), qobject_cast
ensures that the type we cast to actually exist and is in-scope.

We recently noticed that we still have checks for classes that were
removed for Qt 5.0, showing the danger of these stringly-typed
interfaces.

This is not a 100% equivalent construct, since the old code failed
when the checked-for types were in the middle of the dynamic type
hierarchy, while the qobject_cast will succeed. Arguably, the new code
is "correcter" in this regard, but it doesn't matter, since neither of
the two types checked are user-settable.

Amends abbdb4d98d13e78cf47cca3c1d6a049770c57750.

Pick-to: 6.8 6.5
Change-Id: I9828d23f887ac55d29dd83062ac64b8683fa295f
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 5c550afd793efc735d1d37969d19234e8f957fc4)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>